Added styles
The styles for the components and the pages are simple, just to show cleanly the title of the commit and the description (if applied).

Another important thing added to the styles was the tabIndex, which is only for presentation purposes and is not intended to be compliant with the accessibility specification.